Rename to pyrenderdrr
