+++ deepdrr/pyrenderdrr/shaders/dual_peeling_init_vertex.glsl
