+++ deepdrr/pyrenderdrr/shaders/segmentation.vert
